Extract legacy page code into a file
